+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a bug where request builder classes for collections endpoints would have a wrong name. #1052
 - Fixed a bug where symbols (properties, methods, classes) could contain invalid characters #1436
+- Renamed parameters for requests: o => Options, q => queryParameters, h => headers. [#1380](https://github.com/microsoft/kiota/issues/1380)
 
 ## [0.0.20] - 2022-03-25